fix(logsearch): 修复日志数据量过大引起数量统计溢出

+++ miaocha-server/src/main/java/com/hinadt/miaocha/application/service/impl/logsearch/executor/DetailSearchExecutor.java
@@ -69,7 +69,7 @@ public class DetailSearchExecutor extends BaseSearchExecutor
             QueryResult countQueryResult = countFuture.get();
 
             resultProcessor.processDetailResult(detailQueryResult, result);
-            int totalCount = resultProcessor.processTotalCountResult(countQueryResult);
+            long totalCount = resultProcessor.processTotalCountResult(countQueryResult);
             result.setTotalCount(totalCount);
 
             return result;
